refactor(kubernetes): Move ManifestController to Kubernetes
The ManifestController is currently in clouddriver-web even though
it only applies to Kubernetes accounts; this is causing a number
of kubernetes-specific classes to live in clouddriver-core.

Move the ManifestController to live in clouddriver-kubernetes-v2,
and also move the classes it depdends on out of clouddriver-core
and into clouddriver-kubernetes-v2.
