+++ clouddriver-kubernetes-v2/src/main/java/com/netflix/spinnaker/config/KubernetesV2Configuration.java
@@ -23,8 +23,11 @@ import com.netflix.spinnaker.clouddriver.kubernetes.v2.caching.KubernetesV2Provi
 import com.netflix.spinnaker.clouddriver.kubernetes.v2.caching.agent.KubernetesV2CachingAgentDispatcher;
 import com.netflix.spinnaker.clouddriver.kubernetes.v2.description.manifest.GlobalKubernetesKindRegistry;
 import com.netflix.spinnaker.clouddriver.kubernetes.v2.description.manifest.KubernetesKindProperties;
+import com.netflix.spinnaker.clouddriver.kubernetes.v2.model.ManifestProvider;
+import com.netflix.spinnaker.clouddriver.kubernetes.v2.model.NoopManifestProvider;
 import com.netflix.spinnaker.clouddriver.kubernetes.v2.security.KubernetesV2Credentials;
 import com.netflix.spinnaker.clouddriver.security.AccountCredentialsRepository;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
@@ -63,5 +66,11 @@ public class KubernetesV2Configuration {
         kubernetesConfigurationProperties,
         credentialFactory,
         catsModule);
+  }
+
+  @Bean
+  @ConditionalOnMissingBean(ManifestProvider.class)
+  public ManifestProvider noopManifestProvider() {
+    return new NoopManifestProvider();
   }
 }
